+++ src/protontricks/__init__.py
@@ -1,6 +1,7 @@
 from .steam import *
 from .winetricks import *
 from .gui import *
+from .util import *
 
 from ._version import get_versions
 __version__ = get_versions()['version']