+++ src/protontricks/winetricks.py
@@ -4,9 +4,9 @@ import shutil
 import subprocess
 import sys
 
-__all__ = (
-    "get_winetricks_path", "run_winetricks_command"
-)
+from protontricks.util import run_command
+
+__all__ = ("get_winetricks_path",)
 
 logger = logging.getLogger("protontricks")
 
@@ -40,45 +40,3 @@ def get_winetricks_path():
         "'winetricks' executable could not be found automatically."
     )
     return None
-
-
-def run_winetricks_command(
-        steam_path, winetricks_path, proton_app, steam_app, command):
-    """Run a Winetricks command inside the Wine prefix for the selected
-    Proton-enabled Steam app.
-
-    The environment variables are changed for the duration of the
-    call and restored afterwards.
-    """
-    # Make a copy of the environment variables to restore later
-    environ_copy = os.environ.copy()
-
-    if not os.environ.get("WINE"):
-        logger.info(
-            "WINE environment variable is not available. "
-            "Setting WINE environment variable to Proton bundled version"
-        )
-        os.environ["WINE"] = os.path.join(
-            proton_app.install_path, "dist", "bin", "wine")
-
-    if not os.environ.get("WINESERVER"):
-        logger.info(
-            "WINESERVER environment variable is not available. "
-            "Setting WINESERVER environment variable to Proton bundled version"
-        )
-        os.environ["WINESERVER"] = os.path.join(
-            proton_app.install_path, "dist", "bin", "wineserver"
-        )
-
-    os.environ["WINETRICKS"] = winetricks_path
-    os.environ["WINEPREFIX"] = steam_app.prefix_path
-
-    # Unset WINEARCH, which might be set for another Wine installation
-    os.environ.pop("WINEARCH", "")
-
-    try:
-        subprocess.call([winetricks_path] + command)
-    finally:
-        # Restore original env vars
-        os.environ.clear()
-        os.environ.update(environ_copy)